refactor(pdf-service): move attachment logic to shared domains folder
