+++ app/domains/shared/pdf/attachment/createAttachmentPages.ts
@@ -1,7 +1,7 @@
 import type PDFDocument from "pdfkit";
 import type { BeratungshilfeFormularContext } from "~/domains/beratungshilfe/formular";
 import type { ProzesskostenhilfeFormularContext } from "~/domains/prozesskostenhilfe/formular";
-import { createAttachmentEntries } from "~/domains/prozesskostenhilfe/services/pdf/attachment/createAttachmentEntries";
+import { createAttachmentEntries } from "~/domains/shared/pdf/attachment/createAttachmentEntries";
 import type { AttachmentEntries } from "~/services/pdf/attachment";
 import { createHeading } from "~/services/pdf/createHeading";
 import { createHeader } from "~/services/pdf/header/createHeader";
